Update transformation-convert-to-csv-lambdaFunction
In this data engineering project, we will explore the creation and automation of a Python ETL workflow. This workflow involves extracting real estate property data from the Zillow Rapid API, storing it in an Amazon S3 bucket, and subsequently triggering a sequence of Lambda functions for data transformation. The data will be converted into CSV format and loaded into another S3 bucket, all orchestrated by Apache Airflow.

To ensure a seamless process, Apache Airflow will utilize an S3KeySensor operator to monitor the availability of the transformed data in an AWS S3 bucket before attempting to load it into an Amazon Redshift data warehouse. Once this is accomplished, we will connect Amazon QuickSight to the Redshift cluster for visualizing the Zillow Rapid data. This project will be executed entirely on the AWS cloud platform.